# your_project/docs/swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
  CompaniesDetails:
    description: CompaniesDetails represents detailed information about a specific
      company.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the company entry
          was created.
        type: string
      email:
        description: Email is the email address associated with the company.
        type: string
      id:
        description: ID is the unique identifier for the company.
        type: string
      name:
        description: Name is the name of the company.
        type: string
      website:
        description: Website is the website URL of the company.
        type: string
    type: object
  CompaniesPagination:
    description: CompaniesPagination represents the paginated list of companies.
    properties:
      items:
        description: Items is a slice containing individual company details.
        items:
          $ref: '#/definitions/CompaniesTable'
        type: array
      limit:
        description: Limit is the maximum number of items per page in the pagination.
        type: integer
      page:
        description: Page is the current page number in the pagination.
        type: integer
      totalCount:
        description: TotalCount is the total number of companies in the entire list.
        type: integer
    type: object
  CompaniesTable:
    description: CompaniesTable represents a single company entry in a table.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the company entry
          was created.
        type: string
      email:
        description: Email is the email address associated with the company.
        type: string
      id:
        description: ID is the unique identifier for the company.
        type: string
      name:
        description: Name is the name of the company.
        type: string
    type: object
  CompanyIn:
    description: MailinglistIn represents the input structure for creating a new mailinglist.
    properties:
      description:
        description: Description is the description of the mailinglist.
        type: string
      name:
        description: Name is the name of the mailinglist.
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /companies:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CompaniesPagination'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      parameters:
      - description: Company query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CompanyIn'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
  /companies/{ID}:
    delete:
      parameters:
      - description: Company ID
        in: path
        name: ID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
    get:
      parameters:
      - description: Company ID
        in: path
        name: ID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CompaniesDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      parameters:
      - description: Company ID
        in: path
        name: ID
        required: true
        type: string
      - description: Company query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CompanyIn'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"